tests/runtests.sh: Add LOCALDIR environment variable
If you have LOCALDIR set in your environment then when you execute
runtests.sh it uses LOCALDIR for the toaster workdir and runs the
container with --local set.  This requires that LOCALDIR has a poky
directory in it.  The primary use case for this is so that you can run
the tests against a local checkout of poky that you are actively
developing.

+++ tests/runtests.sh
@@ -46,6 +46,11 @@
 # SELENIUM_TIMEOUT
 #    The amount of time to wait for elements in selenium. The default will
 #    be the default for the --timeout argument to smoketests, which is 120.
+#
+# LOCALDIR
+#    If this is set, the dir is used as a workdir, poky is assumed to be there,
+#    and the toaster container is run in --local mode.  This is intended to allow
+#    you to run the test on the local checkout of poky you are working on.
 
 set -e
 
@@ -90,7 +95,7 @@ function fail () {
 }
 
 function start_toaster() {
-    mkdir $tempdir/toasterbuild
+    mkdir -p $tempdir/$toasterbuild
 
     touch $toasterlog
     sentinel="Successful start."
@@ -102,8 +107,8 @@ function start_toaster() {
 
     printf "\n\nStarting toaster...\n"
     docker run -t --rm=true --name=$toastername \
-               -v $tempdir/toasterbuild:/workdir \
-               $image >> $toasterlog 2>&1 &
+               -v $tempdir/$toasterbuild:/workdir \
+               ${image} ${local_arg} >> $toasterlog 2>&1 &
     toasterpid=$!
 
     while ! grep "$sentinel" $toasterlog >& /dev/null; do
@@ -149,7 +154,17 @@ function start_selenium() {
     done
 }
 
-tempdir=$(mktemp -d --suffix toastertest --tmpdir)
+if [ "" != "$LOCALDIR" ]; then
+    tempdir="$LOCALDIR"
+    # if we are running locally, we want everything in
+    # our workdir.
+    toasterbuild=""
+    local_arg="--local"
+else
+    tempdir=$(mktemp -d --suffix toastertest --tmpdir)
+    toasterbuild="toasterbuild"
+    local_arg=""
+fi
 echo "Logs in $tempdir"
 
 id=$(uuidgen)
@@ -158,7 +173,7 @@ seleniumname=seleniumserver-$id
 
 toasterlog=$tempdir/toaster.log
 seleniumlog=$tempdir/selenium.log
-toaster_ui_log=$tempdir/toasterbuild/build-toaster-2/toaster_ui.log
+toaster_ui_log=$tempdir/$toasterbuild/build-toaster-2/toaster_ui.log
 
 if [ "" != "$IMAGE" ]; then
     image="$IMAGE"
@@ -207,13 +222,24 @@ ${SCRIPT_DIR}/smoketests.py --toaster_url="$toastername:8000" \
                 $timeout_arg \
                 $pokybranch_arg
 echo "smoketests PASSED!"
-${SCRIPT_DIR}/checkartifacts.sh $tempdir/toasterbuild/build-toaster-2
+
+${SCRIPT_DIR}/checkartifacts.sh $tempdir/$toasterbuild/build-toaster-2
 echo "checkartifacts PASSED!"
 
-grep -q "No migrations to apply" $toasterlog
-echo "no migration PASSED"
+grep -q "No migrations to apply" $toasterlog || test "" != $LOCALDIR
+echo "no migration priming PASSED"
 
+if grep -q "INFO Fetching layers" $toasterlog; then
+    if [ "" == $LOCALDIR ]; then
+	exit 1
+    fi
+fi
+echo "layerindex priming PASSED"
 echo "ALL TESTS PASSED!"
 
 stop_containers
-rm $tempdir -rf
+if [ "" == "$LOCALDIR" ]; then
+    # tempdir is our workdir, so do NOT
+    # rm it if we are running locally
+    rm $tempdir -rf
+fi
